Replace use of MPI shared memory windows with POSIX shared memory.
The MPIShared class was already using MPI shared windows in a
synchronous fashion with explicit broadcast of data during set()
operations.  This replaces node-local MPI shared memory with basic
POSIX shared memory segments.  This works because were not doing
RMA operations on this memory anyway.  The MPILock class does make
use of one-sided RMA operations to implement locking, and that
behavior remains.  This work adds a new dependency on the posix-ipc
package.